+++ pkg/machinepool/mutate_machinepool_create_test.go
@@ -6,6 +6,8 @@ import (
 	"reflect"
 	"testing"
 
+	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2019-07-01/compute"
+	"github.com/Azure/go-autorest/autorest/to"
 	"github.com/giantswarm/apiextensions/v2/pkg/apis/release/v1alpha1"
 	"github.com/giantswarm/apiextensions/v3/pkg/annotation"
 	"github.com/giantswarm/microerror"
@@ -13,9 +15,11 @@ import (
 	"k8s.io/api/admission/v1beta1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
+	capzv1alpha3 "sigs.k8s.io/cluster-api-provider-azure/exp/api/v1alpha3"
 	"sigs.k8s.io/cluster-api/api/v1alpha3"
 
 	builder "github.com/giantswarm/azure-admission-controller/internal/test/machinepool"
+	"github.com/giantswarm/azure-admission-controller/internal/vmcapabilities"
 	"github.com/giantswarm/azure-admission-controller/pkg/mutator"
 	"github.com/giantswarm/azure-admission-controller/pkg/unittest"
 )
@@ -154,6 +158,39 @@ func TestMachinePoolCreateMutate(t *testing.T) {
 			},
 			errorMatcher: nil,
 		},
+		{
+			name:     fmt.Sprintf("case 5: auto selection of failure domains with valid number requested"),
+			nodePool: builder.BuildMachinePoolAsJson(builder.Replicas(1), builder.Annotation(annotationAutoFailureDomain, "3")),
+			patches: []mutator.PatchOperation{
+				{
+					Operation: "add",
+					Path:      "/spec/failureDomains",
+					Value:     []string{"1", "2", "3"},
+				},
+				{
+					Operation: "remove",
+					Path:      "/metadata/annotations/giantswarm.io~1auto-availability-zones-count",
+					Value:     nil,
+				},
+				{
+					Operation: "replace",
+					Path:      "/metadata/labels/cluster.x-k8s.io~1cluster-name",
+					Value:     "",
+				},
+				{
+					Operation: "add",
+					Path:      "/spec/minReadySeconds",
+					Value:     float64(0),
+				},
+			},
+			errorMatcher: nil,
+		},
+		{
+			name:         fmt.Sprintf("case 6: auto selection of failure domains with invalid number requested"),
+			nodePool:     builder.BuildMachinePoolAsJson(builder.Replicas(1), builder.Annotation(annotationAutoFailureDomain, "4")),
+			patches:      []mutator.PatchOperation{},
+			errorMatcher: IsTooManyFailureDomainsRequestedError,
+		},
 	}
 
 	for _, tc := range testCases {
@@ -207,9 +244,111 @@ func TestMachinePoolCreateMutate(t *testing.T) {
 				t.Fatal(err)
 			}
 
+			// AzureMachinePool
+			amp := &capzv1alpha3.AzureMachinePool{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "ab123",
+					Namespace: "org-giantswarm",
+					Labels: map[string]string{
+						"azure-operator.giantswarm.io/version": "5.0.0",
+					},
+				},
+				Spec: capzv1alpha3.AzureMachinePoolSpec{
+					Location: "westeurope",
+					Template: capzv1alpha3.AzureMachineTemplate{
+						VMSize: "Standard_A2_v2",
+					},
+				},
+			}
+			err = ctrlClient.Create(ctx, amp)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			stubbedSKUs := map[string]compute.ResourceSku{
+				"Standard_A2_v2": {
+					Name: to.StringPtr("Standard_A2_v2"),
+					Capabilities: &[]compute.ResourceSkuCapabilities{
+						{
+							Name:  to.StringPtr("AcceleratedNetworkingEnabled"),
+							Value: to.StringPtr("False"),
+						},
+						{
+							Name:  to.StringPtr("vCPUs"),
+							Value: to.StringPtr("4"),
+						},
+						{
+							Name:  to.StringPtr("MemoryGB"),
+							Value: to.StringPtr("16"),
+						},
+					},
+					LocationInfo: &[]compute.ResourceSkuLocationInfo{
+						{
+							Location: to.StringPtr("westeurope"),
+							Zones:    &[]string{"1", "2", "3"},
+						},
+					},
+				},
+				"Standard_A4_v2": {
+					Name: to.StringPtr("Standard_A4_v2"),
+					Capabilities: &[]compute.ResourceSkuCapabilities{
+						{
+							Name:  to.StringPtr("AcceleratedNetworkingEnabled"),
+							Value: to.StringPtr("False"),
+						},
+						{
+							Name:  to.StringPtr("vCPUs"),
+							Value: to.StringPtr("4"),
+						},
+						{
+							Name:  to.StringPtr("MemoryGB"),
+							Value: to.StringPtr("16"),
+						},
+					},
+					LocationInfo: &[]compute.ResourceSkuLocationInfo{
+						{
+							Location: to.StringPtr("westeurope"),
+							Zones:    &[]string{"1", "2"},
+						},
+					},
+				},
+				"Standard_A8_v2": {
+					Name: to.StringPtr("Standard_A8_v2"),
+					Capabilities: &[]compute.ResourceSkuCapabilities{
+						{
+							Name:  to.StringPtr("AcceleratedNetworkingEnabled"),
+							Value: to.StringPtr("False"),
+						},
+						{
+							Name:  to.StringPtr("vCPUs"),
+							Value: to.StringPtr("4"),
+						},
+						{
+							Name:  to.StringPtr("MemoryGB"),
+							Value: to.StringPtr("16"),
+						},
+					},
+					LocationInfo: &[]compute.ResourceSkuLocationInfo{
+						{
+							Location: to.StringPtr("westeurope"),
+							Zones:    &[]string{},
+						},
+					},
+				},
+			}
+			stubAPI := NewStubAPI(stubbedSKUs)
+			vmcaps, err := vmcapabilities.New(vmcapabilities.Config{
+				Azure:  stubAPI,
+				Logger: newLogger,
+			})
+			if err != nil {
+				panic(microerror.JSON(err))
+			}
+
 			admit, err := NewCreateMutator(CreateMutatorConfig{
 				CtrlClient: ctrlClient,
 				Logger:     newLogger,
+				VMcaps:     vmcaps,
 			})
 			if err != nil {
 				t.Fatal(err)